Backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-core
langchain-groq
groq
langgraph
langsmith
cryptography
pymongo